requirements.txt
tensorflow-gpu==1.8.0
matplotlib==2.1.2

